scale back try_write function on unix

+++ conda/utils.py
@@ -1,15 +1,15 @@
 from __future__ import print_function, division, absolute_import
 
-import logging
-import sys
-import hashlib
 import collections
-from functools import partial
-from os.path import abspath, isdir, join
+import hashlib
+import logging
 import os
 import re
 import subprocess
+import sys
 import tempfile
+from functools import partial
+from os.path import abspath, isdir, join, isfile
 
 
 log = logging.getLogger(__name__)
@@ -89,18 +89,35 @@ def gnu_get_libc_version():
     return f()
 
 
-def try_write(dir_path):
+def try_write(dir_path, heavy=False):
+    """Test write access to a directory.
+
+    Args:
+        dir_path (str): directory to test write access
+        heavy (bool): Actually create and delete a file, or do a faster os.access test.
+           https://docs.python.org/dev/library/os.html?highlight=xattr#os.access
+
+    Returns:
+        bool
+
+    """
     if not isdir(dir_path):
         return False
-    # try to create a file to see if `dir_path` is writable, see #2151
-    temp_filename = join(dir_path, '.conda-try-write-%d' % os.getpid())
-    try:
-        with open(temp_filename, mode='wb') as fo:
-            fo.write(b'This is a test file.\n')
-        os.unlink(temp_filename)
-        return True
-    except (IOError, OSError):
-        return False
+    if on_win or heavy:
+        # try to create a file to see if `dir_path` is writable, see #2151
+        temp_filename = join(dir_path, '.conda-try-write-%d' % os.getpid())
+        try:
+            with open(temp_filename, mode='wb') as fo:
+                fo.write(b'This is a test file.\n')
+            os.unlink(temp_filename)
+            return True
+        except (IOError, OSError):
+            return False
+        finally:
+            if isfile(temp_filename):
+                os.unlink(temp_filename)
+    else:
+        return os.access(dir_path, os.W_OK)
 
 
 def hashsum_file(path, mode='md5'):
